Fix property sheet error from not finding labelMorph font When getting the property sheet of a labelMorph that is using a font that doesn't exist in the 'x11Globals scalableFont' list an error occurs and the property sheet can't display.
The scalableFont list has an undefinedSelector method which handles
direct requests for fonts that don't exist and returns 'times'.
Unfortunately the enumeration of the fonts that the property sheet
uses just lists the selectors of 'fontFamily' so treats the
undefinedSelect method as a font name and hence the default handlin
never occurs.

This fix taken here adds a 'defaultFont' to the scalableFont list.
This returns 'times', the same as the undefined selector. The
property sheet looks this up if the actual font name can't be found.

I also remove the undefined selector slot name and a couple of other
non-font slots that were missed in the x11Globals fontFamily method.